feat: disable filter buttons that would result in no matches
Or well, at least partially. When the user is only using the filter buttons, this works fine. As soon as they start using the search input as well, they can still end up in situations with 0 matches. We can prevent that by running a presearch of `${query} ${tag}` for each visible tag, but that impacts performance quite drastically.