feat: add recommendation test code snippet
Hi @kentcdodds,

Thanks for all of your job. I really apreciate it and It makes my job easier. 

I recently was searching about query priority and I found this article. When i was reading it, I realize that could be nicer if you show to the lector a code snippet of your test recommendation. 

What do you think?