Update testing-implementation-details.mdx
I propose this change because the functionality to have multiple accordion items to be open at once is not working as I think it should. I am not sure if the line: tsx remove=2-3,16,25 add=4-14,17,26 if right, as it if diffucult to check without have the visual help. I hope this make the content more clear. Good post by the way. Thanks for sharing.

P.S. this is the first time I propose change to this project.