enable MDX type checking
In `tsconfig.json` `**/*.mdx` is now included explicitly. This makes the
MDX files part of the same TypeScript program as the TypeScript files.
This is only needed at all, because the `include` array is defined. It
can also be omitted entirely.

The `mdx.checkMdx` option in `tsconfig.json` enables strict type
checking for MDX files. It’s analogous to `checkJs` for JavaScript
files.

The global type `MDXProvidedComponents` specified which types are
available inside MDX files globally.